# swagger.json
{
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "schemes": [
    "https"
  ],
  "swagger": "2.0",
  "info": {
    "description": "Aplikacija nudi podporo pri popisu vrst po Sloveniji",
    "title": "Biolog API.",
    "contact": {
      "name": "David Rubin",
      "email": "david.rubin95@gmail.com"
    },
    "version": "1.0.0"
  },
  "host": "localhost:4000",
  "basePath": "/api/v1",
  "paths": {
    "/auth_providers": {
      "get": {
        "description": "Pridobi vse mozne ponudnike avtentikacije",
        "tags": [
          "authproviders"
        ],
        "operationId": "getAuthProviders",
        "responses": {
          "200": {
            "$ref": "#/responses/authProviderResponse"
          },
          "400": {
            "$ref": "#/responses/"
          }
        }
      }
    },
    "/auth_providers/{id}": {
      "get": {
        "description": "Pridobi podrobnosti o dolocenem ponudniku avtentikacije, ki je pri nas na voljo",
        "tags": [
          "authproviders"
        ],
        "operationId": "getAuthProviderByID",
        "parameters": [
          {
            "minimum": 1,
            "type": "integer",
            "format": "int64",
            "description": "ID ponudnika avtentikacije",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/responses/authProviderResponse"
          },
          "400": {
            "$ref": "#/responses/"
          }
        }
      }
    },
...
    "/users": {
      "get": {
        "description": "Pridobi vse uporabnike, ki imajo vsaj 1 javno opazanje",
        "tags": [
          "users"
        ],
        "operationId": "getUsers",
        "responses": {
          "200": {
            "description": "user",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/user"
              }
            }
          },
          "400": {
            "$ref": "#/responses/"
          }
        }
      },
      "post": {
        "description": "Ustvari novega uporabnika v podatkovni bazi",
        "tags": [
          "users"
        ],
        "operationId": "createUser",
        "parameters": [
          {
            "description": "Podatki o uporabniku'",
            "name": "User",
            "in": "body",
            "required": true,
            "schema": {
              "$ref": "#/definitions/user"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "user",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "400": {
            "$ref": "#/responses/"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "description": "Pridobi podrobnosti o uporabniku",
        "tags": [
          "users"
        ],
        "operationId": "getUserByID",
        "parameters": [
          {
            "maximum": 99999999,
            "minimum": 10000000,
            "type": "integer",
            "format": "int32",
            "description": "ID uporabnika",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "user",
            "schema": {
              "$ref": "#/definitions/user"
            }
          },
          "400": {
            "$ref": "#/responses/"
          }
        }
      },
      "delete": {
        "description": "Zbrise uproabniski racun in vse zapise o uporabniku",
        "tags": [
          "users"
        ],
        "operationId": "deleteUser",
        "parameters": [
          {
            "maximum": 99999999,
            "minimum": 10000000,
            "type": "integer",
            "format": "int32",
            "description": "ID uporabnika",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {},
          "400": {
            "$ref": "#/responses/"
          }
        }
      },
      "patch": {
        "description": "Posodobi podatke o uporabniku",
        "tags": [
          "users"
        ],
        "operationId": "updateUser",
        "parameters": [
          {
            "maximum": 99999999,
            "minimum": 10000000,
            "type": "integer",
            "format": "int32",
            "description": "ID uporabnika",
            "name": "ID",
            "in": "path",
            "required": true
          }
        ],
        "responses": {
          "204": {},
          "400": {
            "$ref": "#/responses/"
          }
        }
      }
    }
  },
  "definitions": {
    "authProvider": {
      "description": "Placeholder, ce bi uporabili vec avtentikatorjev (recimo Google + Facebook)",
      "type": "object",
      "title": "AuthProvider (zunanji avtentikator)",
      "required": [
        "ID",
        "Name"
      ],
      "properties": {
        "ID": {
          "description": "Identifikator ponudnika avtentikacije",
          "type": "integer",
          "format": "int64",
          "example": 1
        },
        "Name": {
          "description": "Ime ponudnika avtentikacije",
          "type": "string",
          "maxLength": 32,
          "example": "Google"
        }
      },
      "x-go-name": "AuthProvider",
      "x-go-package": "github.com/rubinda/biolog"
    },
...
    "user": {
      "description": "Podatki o uporabniku so vzeti od zunanjega avtentikatorja",
      "type": "object",
      "title": "User (uporabnik nase aplikacijo)",
      "required": [
        "id",
        "displayName",
        "externalID",
        "givenName",
        "familyName",
        "email"
      ],
      "properties": {
        "displayName": {
          "description": "Prikazno ime za uporabnika",
          "type": "string",
          "maxLength": 64,
          "x-go-name": "DisplayName",
          "example": "David Rubin"
        },
        "email": {
          "description": "Elektronski naslov uporabnika",
          "type": "string",
          "format": "email",
          "maxLength": 128,
          "x-go-name": "Email"
        },
        "externalID": {
          "description": "ID zunanjega avtentikatorja za tega uporabnika",
          "type": "string",
          "maxLength": 255,
          "x-go-name": "ExternalID",
          "example": "8457232358972358923566"
        },
        "familyName": {
          "description": "Priimek uporabnika",
          "type": "string",
          "maxLength": 32,
          "pattern": "[A-Za-z]+",
          "x-go-name": "FamilyName",
          "example": "Rubin"
        },
        "givenName": {
          "description": "Ime uporabnika",
          "type": "string",
          "maxLength": 32,
          "pattern": "[A-Za-z]+",
          "x-go-name": "GivenName",
          "example": "David"
        },
        "id": {
          "description": "8 mestni ID dolocenega uporabnika",
          "type": "integer",
          "format": "int64",
          "maximum": 99999999,
          "minimum": 10000000,
          "x-go-name": "ID",
          "example": 10000000
        },
        "picture": {
          "description": "URL do prikazne slike uporabnika",
          "type": "string",
          "maxLength": 255,
          "x-go-name": "Picture",
          "example": "https://www.biologapp.com/static/img/david-profile.png"
        },
        "publicObservations": {
          "description": "Pove ali uporabnik dovoli javen dostop do svojih opazanj",
          "type": "boolean",
          "x-go-name": "PublicObservations"
        }
      },
      "x-go-name": "User",
      "x-go-package": "github.com/rubinda/biolog"
    }
  },
  "responses": {
    "authProviderResponse": {
      "description": "AuthProviderResponse model.\n\nSe uporablja pri virih, ki vracajo podatke o zunanjih avtentikatorjih",
      "schema": {
        "$ref": "#/definitions/authProvider"
      }
    }
  }
}